fix(tests): remove line failing on power-pc
It's worth noting that I never reproduced the issue, nor have I seen
a stack-trace. Thus the line is removed in good-faith, but should also
not pose any problem considering it was very specific and only in a
test-case.

The main problem is that I don't understand anymore why that assertion
should be true, and thus can't judge the correctness of this fix at all.

Closes #25