Update ci, in line with gitdb
This updates smmap's CI configuration in ways that are in line with
recent updates to gitdb's. In most cases there is no difference in
the changes, and the reason for the updates is more to avoid
confusing differences than from the value of the changes
themselves. In one case, there is a major difference (fetch-depth).

- https://github.com/gitpython-developers/gitdb/pull/89 (same)

- https://github.com/gitpython-developers/gitdb/pull/90 (same)
  It's just the project, not dependencies, but otherwise the same.

- https://github.com/gitpython-developers/gitdb/pull/92 (opposite)
  This is the major difference. We don't need more than the tip of
  the branch in these tests. Keeping the default fetch-depth of 1
  by not setting it explicitly avoids giving the impression that
  the tests here are doing something they are not (and also serves
  as a speed optimization).

- https://github.com/gitpython-developers/gitdb/pull/93 (same)